+++ CHANGELOG.md
@@ -9,6 +9,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Iterable mixin
+
+#### [Queue](src/linear/queue.ts)
+
+- Add the following methods:
+  - `peak(): T` - Retrieve the first element from the queue, the next to be removed, or throws `QueueUnderflow` exception if is empty
+  - `clear` - remove all elements from the collection
+
+#### [Stack](src/linear/stack.ts)
+
+- Add the following methods:
+
+  - ` clear(): this` - remove all elements from the collection
+  - `[Symbol.iterator]` - iterator
+  - `entries(): IterableIterator<[number, T]>` - Returns an iterable of key, value pairs for every entry in collection
+  - `keys(): IterableIterator<number>` - Returns an iterable of keys in the collection
+  - `values(): IterableIterator<T>` - Returns an iterable of values in the collection
+
 - Add [**LinkedList**](src/list/list.ts) data structure with no limit of size. It has also some exceptions that can be throwed
 - Add methods to [**LinkedList**](src/list/list.ts): `pop`, `push`, `size`, `isEmpty`, `length`, `size`, `indexOf`, `toArray`, `dequeue`, `clear`, `peek`, `tail` and `contains`.
 - Besides the methods above. Add `values`, `entries` (as the default) and `values` iterators to **LinkedList** the same as the [`Array`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/array) has